+++ src/main/java/co/orange/ddanzi/dto/home/HomeResponseDto.java
@@ -1,5 +1,6 @@
 package co.orange.ddanzi.dto.home;
 
+import co.orange.ddanzi.dto.ProductInfo;
 import lombok.Builder;
 import lombok.Getter;
