+++ src/main/java/co/orange/ddanzi/service/HomeService.java
@@ -1,6 +1,7 @@
 package co.orange.ddanzi.service;
 
 import co.orange.ddanzi.domain.product.*;
+import co.orange.ddanzi.dto.ProductInfo;
 import co.orange.ddanzi.dto.home.*;
 import co.orange.ddanzi.global.common.exception.Error;
 import co.orange.ddanzi.global.common.response.ApiResponse;
